chore: Consolidate dotnet new Templates into single class
